refactor: extract performance returns polling
