+++ quality/baseline_report.md
@@ -10,7 +10,7 @@ documentation posture, and validation gates. It is not a completion claim.
 
 ## Code Size Baseline
 
-- Python source files under `src/`: 182
+- Python source files under `src/`: 183
 - Python test files under `tests/`: 93
 - Python packages under `src/`: 9
 - API entry point route/middleware/handler decorators in `src/app/main.py`: 0
@@ -26,7 +26,7 @@ documentation posture, and validation gates. It is not a completion claim.
 | tests/unit/test_attribution_mode_adapter.py | 600 | 21872 |
 | tests/integration/test_risk_calculate.py | 596 | 25093 |
 | tests/integration/test_rolling_metrics_endpoint.py | 566 | 22142 |
-| tests/unit/test_lotus_performance_client.py | 530 | 19686 |
+| tests/unit/test_lotus_performance_client.py | 536 | 19752 |
 | tests/unit/test_rolling_mode_adapter.py | 444 | 16926 |
 | tests/unit/test_drawdown_engine.py | 422 | 16234 |
 | src/app/services/attribution_decomposition.py | 420 | 13767 |
@@ -38,7 +38,7 @@ documentation posture, and validation gates. It is not a completion claim.
 | tests/unit/test_risk_engine_branch_coverage.py | 353 | 13642 |
 | tests/unit/test_concentration_engine_modes.py | 334 | 12149 |
 | src/app/services/calculation_supportability.py | 328 | 11177 |
-| src/app/integrations/lotus_performance_client.py | 326 | 11128 |
+| src/app/services/concentration/parsing.py | 319 | 11074 |
 
 ### Largest Functions And Classes
 
@@ -89,12 +89,12 @@ All checks passed!
 - Ruff format check: passed
 
 ```text
-289 files already formatted
+290 files already formatted
 ```
 - Type checking: passed
 
 ```text
-Success: no issues found in 275 source files
+Success: no issues found in 276 source files
 ```
 - Unit coverage snapshot: passed
 
@@ -106,12 +106,12 @@ Success: no issues found in 275 source files
 ........................................................................ [ 75%]
 ........................................................................ [ 91%]
 ...
-src\app\upstream_errors.py                                 55      0     20      3    96%   181->194, 187->189, 192->194
----------------------------------------------------------------------------------------------------
-TOTAL                                                    5168     80    978     79    97%
+src\app\upstream_errors.py                                    55      0     20      3    96%   181->194, 187->189, 192->194
+------------------------------------------------------------------------------------------------------
+TOTAL                                                       5173     80    978     79    97%
 
-143 files skipped due to complete coverage.
-474 passed in 7.74s
+144 files skipped due to complete coverage.
+474 passed in 8.32s
 ```
 
 ## Complexity And Maintainability Snapshot
@@ -141,7 +141,7 @@ src\app\services\risk\__init__.py - A (100.00)
 - Dependency hygiene: passed
 
 ```text
-Scanning 182 files...
+Scanning 183 files...
 
 Success! No dependency issues found.
 ```
@@ -240,7 +240,7 @@ tests/unit/test_upstream_errors.py::test_classify_upstream_transport_error_matri
 tests/unit/test_upstream_errors.py::test_invalid_payload_and_missing_data_carry_structured_categories
 tests/unit/test_upstream_errors.py::test_extract_upstream_error_detail_variants
 
-474 tests collected in 1.55s
+474 tests collected in 1.61s
 ```
 - Import-linter report-only: passed
